fix: add console log to debug question creation
This commit adds an `onClick` handler with a `console.log` statement to the submit buttons in the organization and volunteer question management pages. This is a debugging measure to verify whether the button click event is being registered correctly, as there is currently no response when adding a new question. This will help to narrow down the source of the problem.