debug: add console logs to handleAddQuestion
This commit adds `console.log` statements to the beginning of the `handleAddQuestion` function in both organization and volunteer question management pages. This is a debugging measure to verify if the function is being called when the form is submitted, as the user is reporting no console output upon clicking the add button.